+++ components/list/PropsType.tsx
@@ -5,6 +5,7 @@ export interface ListPropsType {
   renderHeader?: (() => React.ReactType) | string | JSX.Element;
   renderFooter?: (() => React.ReactType) | string | JSX.Element;
   children?: false | JSX.Element | JSX.Element[];
+  noBorder: boolean;
 }
 
 export interface ListItemPropsType {