+++ components/list/index.tsx
@@ -19,6 +19,7 @@ export default class List extends React.Component<ListProps, any> {
       renderHeader,
       renderFooter,
       styles,
+      noBorder,
       ...restProps
     } = this.props;
 
@@ -52,9 +53,9 @@ export default class List extends React.Component<ListProps, any> {
           return (
             <View {...restProps as any} style={style}>
               {headerDom}
-              <View style={s.Body}>
+              <View style={[s.Body, noBorder && {borderTopWidth:0}]}>
                 {children ? children : null}
-                <View style={[s.BodyBottomLine as ViewStyle]} />
+                <View style={[s.BodyBottomLine as ViewStyle,noBorder && {borderBottomWidth:0}]} />
               </View>
               {footerDom}
             </View>