+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   - textAlign 增加 'right' 选项
   - 把 InputItem clear icon 设置为 x3s
   - InputItem clear 增加 textAlign 值不等于 right 才显示的条件
+- [List](./components/list/index.tsx)
+- 增加了 noBorder 属性
 - [List.Item](./components/list/ListItem.tsx)
   - 增加 `last?: boolean` 属性
 - [Icon](./components/icon/index.tsx)